add info to docs

+++ cleanlab/internal/multiannotator_utils.py
@@ -43,7 +43,7 @@ def assert_valid_inputs_multiannotator(
             )
 
     # Raise error if labels are not formatted properly
-    all_labels_flatten = labels_multiannotator.replace({pd.NA: np.NaN}).values.ravel()
+    all_labels_flatten = labels_multiannotator.replace({pd.NA: np.NaN}).astype(float).values.ravel()
     all_labels_flatten = all_labels_flatten[~np.isnan(all_labels_flatten)]
     assert_valid_class_labels(all_labels_flatten)
